Move Datelike and Timelike traits into a traits module
